Found the actual working code. Also death of a project.
So after rummaging around on my backup drive I think I found the actual
working code I had been working on. The reason this code had not been up
in the first place is because my computer crashed and prior to that I
was still getting used to Github, seeing how it was my first time using
it.